OneTrust_CookieConesnt_Changes: Changed the structure
